請假簽核可以往上送讓下一層看到(1006送給1004)

+++ src/main/java/com/example/fluxeip/model/LeaveRequest.java
@@ -1,5 +1,6 @@
 package com.example.fluxeip.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
@@ -36,7 +37,7 @@ public class LeaveRequest {
     private LocalDateTime endDatetime;
 
     @Column(name = "leave_hours", nullable = false)
-    private Integer leaveHours;
+    private BigDecimal leaveHours;
 
     @Column(name = "reason")
     private String reason;
